Add filter management to Luxid (RAPI and CLI interfaces)
This implements the management (add, remove, replace) of job filter rules
as specified by `doc/design-optables.rst`; it does not yet implement
the filtering logic.

This commit also includes the implementation of filters in the RAPI client.py
and its tests because the client tests check that if all available RAPI
resources are being used by the client, and the only way to make them
be used is to write tests for them.

Signed-off-by: Niklas Hambuechen <niklash@google.com>