JQScheduler: Run scheduler when filters change in config
Until now, job filters being added, modified, or deleted triggered
a scheduler run (`scheduleSomeJobs`) from the place where the filter
modification request is handled (Query.Server).

This could lead to a race:
If the config change containing the updated filters doesn't propagate
fast enough (it is asynchronous), `scheduleSomeJobs` will be triggered
with the old config.

This race only showed up in QAs with larger clusters.

This commit fixes the issue by moving the re-scheduling from the
request-for-config-change location to the place where the
new config is read, so that old and new filters can be compared.

Signed-off-by: Niklas Hambuechen <niklash@google.com>